CHANGE dev page TO debug page
UPDATE UI DEBUG/DEV PAGE
